rename plugin to foreman_reverse_proxy
