+++ app/controllers/foreman_reverse_proxy/hosts_controller.rb
@@ -1,13 +1,13 @@
-module ForemanPluginTemplate
+module ForemanReverseProxy
 
   # Example: Plugin's HostsController inherits from Foreman's HostsController
   class HostsController < ::HostsController
 
     # change layout if needed
-    # layout 'foreman_plugin_template/layouts/new_layout'
+    # layout 'foreman_reverse_proxy/layouts/new_layout'
 
     def new_action
-      # automatically renders view/foreman_plugin_template/hosts/new_action
+      # automatically renders view/foreman_reverse_proxy/hosts/new_action
     end
 
   end
