+++ app/helpers/concerns/foreman_reverse_proxy/hosts_helper_extensions.rb
@@ -1,4 +1,4 @@
-module ForemanPluginTemplate
+module ForemanReverseProxy
   module HostsHelperExtensions
     extend ActiveSupport::Concern
