+++ app/models/concerns/foreman_reverse_proxy/host_extensions.rb
@@ -1,4 +1,4 @@
-module ForemanPluginTemplate
+module ForemanReverseProxy
   module HostExtensions
     extend ActiveSupport::Concern
