+++ app/overrides/dashboard/index/sample_override.html.erb.deface
@@ -1,4 +1,4 @@
 <!-- insert_before 'div:contains("title_action")' -->
-<%= title_actions "<h2>ForemanPluginTemplate</h2>" %>
+<%= title_actions "<h2>ForemanReverseProxy</h2>" %>
 
 <!-- vim: set ft=eruby : -->
