--- app/views/dashboard/_foreman_plugin_template_widget.html.erb
@@ -1,2 +1,0 @@
-<h4 class="header ca"><%= _('ForemanPluginTemplate') %></h4>
-<%= _('Widget content') %>
