+++ app/views/dashboard/_foreman_reverse_proxy_widget.html.erb
@@ -1,0 +1,2 @@
+<h4 class="header ca"><%= _('ForemanReverseProxy') %></h4>
+<%= _('Widget content') %>
