--- app/views/foreman_plugin_template/hosts/new_action.html.erb
@@ -1,1 +1,0 @@
-Welcome to <b>ForemanPluginTemplate</b>
