+++ app/views/foreman_reverse_proxy/hosts/new_action.html.erb
@@ -1,0 +1,1 @@
+Welcome to <b>ForemanReverseProxy</b>
