+++ config/routes.rb
@@ -1,5 +1,5 @@
 Rails.application.routes.draw do
 
-  match 'new_action', :to => 'foreman_plugin_template/hosts#new_action'
+  match 'new_action', :to => 'foreman_reverse_proxy/hosts#new_action'
 
 end
