+++ foreman_reverse_proxy.gemspec
@@ -1,14 +1,14 @@
-require File.expand_path('../lib/foreman_plugin_template/version', __FILE__)
+require File.expand_path('../lib/foreman_reverse_proxy/version', __FILE__)
 
 Gem::Specification.new do |s|
-  s.name        = "foreman_plugin_template"
-  s.version     = ForemanPluginTemplate::VERSION
+  s.name        = "foreman_reverse_proxy"
+  s.version     = ForemanReverseProxy::VERSION
   s.date        = Date.today.to_s
   s.authors     = ["TODO: Your name"]
   s.email       = ["TODO: Your email"]
   s.homepage    = "TODO"
-  s.summary     = "TODO: Summary of ForemanPluginTemplate."
-  s.description = "TODO: Description of ForemanPluginTemplate."
+  s.summary     = "TODO: Summary of ForemanReverseProxy."
+  s.description = "TODO: Description of ForemanReverseProxy."
 
   s.files = Dir["{app,config,db,lib}/**/*"] + ["LICENSE", "Rakefile", "README.md"]
   s.test_files = Dir["test/**/*"]
