--- lib/foreman_plugin_template.rb
@@ -1,4 +1,0 @@
-require "foreman_plugin_template/engine"
-
-module ForemanPluginTemplate
-end
