--- lib/foreman_plugin_template/version.rb
@@ -1,3 +1,0 @@
-module ForemanPluginTemplate
-  VERSION = "0.0.1"
-end
