+++ lib/foreman_reverse_proxy.rb
@@ -1,0 +1,4 @@
+require "foreman_reverse_proxy/engine"
+
+module ForemanReverseProxy
+end
