+++ lib/foreman_reverse_proxy/version.rb
@@ -1,0 +1,3 @@
+module ForemanReverseProxy
+  VERSION = "0.0.1"
+end
