+++ lib/tasks/foreman_reverse_proxy_tasks.rake
@@ -1,5 +1,5 @@
 # Tasks
-namespace :foreman_plugin_template do
+namespace :foreman_reverse_proxy do
   namespace :example do
     desc 'Example Task'
     task :task => :environment do
@@ -10,8 +10,8 @@ end
 
 # Tests
 namespace :test do
-  desc "Test ForemanPluginTemplate"
-  Rake::TestTask.new(:foreman_plugin_template) do |t|
+  desc "Test ForemanReverseProxy"
+  Rake::TestTask.new(:foreman_reverse_proxy) do |t|
     test_dir = File.join(File.dirname(__FILE__), '../..', 'test')
     t.libs << ["test",test_dir]
     t.pattern = "#{test_dir}/**/*_test.rb"
@@ -20,12 +20,12 @@ namespace :test do
 end
 
 Rake::Task[:test].enhance do
-  Rake::Task['test:foreman_plugin_template'].invoke
+  Rake::Task['test:foreman_reverse_proxy'].invoke
 end
 
 load 'tasks/jenkins.rake'
 if Rake::Task.task_defined?(:'jenkins:unit')
   Rake::Task["jenkins:unit"].enhance do
-    Rake::Task['test:foreman_plugin_template'].invoke
+    Rake::Task['test:foreman_reverse_proxy'].invoke
   end
 end
