+++ test/factories/foreman_reverse_proxy_factories.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
   factory :host do
-    name 'foreman_plugin_template'
+    name 'foreman_reverse_proxy'
   end
 end
