+++ test/unit/foreman_reverse_proxy_test.rb
@@ -1,6 +1,6 @@
 require 'test_plugin_helper'
 
-class ForemanPluginTemplateTest < ActiveSupport::TestCase
+class ForemanReverseProxyTest < ActiveSupport::TestCase
   setup do
     User.current = User.find_by_login "admin"
   end
